+++ r_code/image_impl.h
@@ -40,7 +40,9 @@ namespace r_code {
 
 class dll_export ImageImpl {
 private:
-    uintptr_t *_data; // [object map|code segment|object names]
+    /// [object map|code segment|object names]
+    uint32_t *_data;
+
     uint64_t _timestamp;
     size_t _map_size;
     size_t _code_size;
@@ -50,13 +52,13 @@ protected:
     size_t map_size() const;
     size_t code_size() const;
     size_t names_size() const;
-    uintptr_t *data() const;
-    uintptr_t &data(size_t i);
-    uintptr_t &data(size_t i) const;
+    uint32_t *data() const;
+    uint32_t &data(size_t i);
+    uint32_t &data(size_t i) const;
 public:
     void *operator new(size_t, uint64_t data_size);
     void operator delete(void *o);
-    ImageImpl(uint64_t timestamp, uint64_t map_size, uint64_t code_size, uint64_t names_size);
+    ImageImpl(uint64_t timestamp, size_t map_size, size_t code_size, size_t names_size);
     ~ImageImpl();
 };
 }
